Fix Vec3.__mod__ annotations and ignores
* Allow use of float (2.0 % 0.5 works, also % 0.3)

* Add inner ignores

+++ pyglet/math.py
@@ -400,11 +400,19 @@ class Vec3(_typing.NamedTuple):
     def __trunc__(self) -> Vec3:
         return Vec3(_math.trunc(self[0]), _math.trunc(self[1]), _math.trunc(self[2]))
 
-    def __mod__(self, other: Vec3 | tuple[int, int, int] | float) -> Vec3:
+    def __mod__(self, other: Vec3 | tuple[float, float, float] | float) -> Vec3:
         try:
-            return Vec3(self[0] % other[0], self[1] % other[1], self[2] % other[2])
+            return Vec3(
+                self[0] % other[0],  # type: ignore
+                self[1] % other[1],  # type: ignore
+                self[2] % other[2]   # type: ignore
+            )
         except TypeError:
-            return Vec3(self[0] % other, self[1] % other, self[2] % other)
+            return Vec3(
+                 self[0] % other,  # type: ignore
+                 self[1] % other,  # type: ignore
+                 self[2] % other   # type: ignore
+            )
 
     def __pow__(self, other: Vec3 | tuple[float, float, float] | float) -> Vec3:
         try:
